TASK-REMOTE-SETUP: publish sanitized shared project snapshot

+++ migrations/009_supplier_finances.sql
@@ -4,14 +4,14 @@ PRAGMA foreign_keys = ON;
 -- /v2/finances (Rosstat/GIR BO data). Separate table, not columns added to
 -- global_supplier_registry (migration 008) — that table already exists in
 -- deployed databases and SQLite has no "ALTER TABLE ADD COLUMN IF NOT
--- EXISTS"; this repo re-runs every migration on every startup (see
+-- EXISTS" — this repo re-runs every migration on every startup (see
 -- MailRepository.ensure_schema), so schema changes must stay purely
 -- additive (new table) to remain safe to re-run. See
--- docs/suppliers-screen.md §5.1 and PROJECT_DOCUMENTATION.md §2.
+-- Documents/28-8/suppliers-screen.md §5.1 and Documents/28-8/PROJECT_DOCUMENTATION.md §2.
 CREATE TABLE IF NOT EXISTS global_supplier_finances (
     global_supplier_id INTEGER PRIMARY KEY REFERENCES global_suppliers(id) ON DELETE CASCADE,
     report_year INTEGER,
-    revenue INTEGER,
-    profit INTEGER,
+    revenue BIGINT,
+    profit BIGINT,
     updated_at TEXT NOT NULL DEFAULT ''
 );
